django-cloud-hosting/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
